+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
+## Upcoming
+
+- `#4`: As dev, I want to use jestjs for test.
+
 ## `0.2.0`
 
-- As js dev, I want to provide child values as array.
+- `#3`: As web dev, I want to provide child values as array.
 
 ## `0.1.0`
 
-- initial: As javascript developer, I want to have html tags in functions.
+- `#1`: As web dev, I want to have html tags in functions.